feat: rename config and add to vcs. Update CI action
